Switch from `uvu` to Node’s test runner
